React playground reusable
