+++ packages/playground/src/main.tsx
@@ -1,7 +1,11 @@
 import editorWorker from "monaco-editor/esm/vs/editor/editor.worker?worker";
 import jsonWorker from "monaco-editor/esm/vs/language/json/json.worker?worker";
 import { createRoot } from "react-dom/client";
-import { App } from "./app";
+import { FunctionComponent } from "react";
+import { createBrowserHost } from "./browser-host";
+import { Playground } from "./components/cadl-playground";
+import { attachServices } from "./services";
+
 import "./style.css";
 
 (self as any).MonacoEnvironment = {
@@ -11,6 +15,14 @@ import "./style.css";
     }
     return new editorWorker();
   },
+};
+
+
+const host = await createBrowserHost();
+await attachServices(host);
+
+ const App: FunctionComponent = () => {
+  return <Playground host={host} />;
 };
 
 const root = createRoot(document.getElementById("root")!);
